Optimise day 15 - use int32

+++ day15_rambunctious_recitation/main_test.go
@@ -5,9 +5,9 @@ import (
 )
 
 func TestPart1(t *testing.T) {
-	input := []int{1, 20, 8, 12, 0, 14}
+	input := []int32{1, 20, 8, 12, 0, 14}
 
-	want := 492
+	want := int32(492)
 	got := solve(input, 2020)
 
 	if got != want {
@@ -16,9 +16,9 @@ func TestPart1(t *testing.T) {
 }
 
 func TestPart2(t *testing.T) {
-	input := []int{1, 20, 8, 12, 0, 14}
+	input := []int32{1, 20, 8, 12, 0, 14}
 
-	want := 63644
+	want := int32(63644)
 	got := solve(input, 30000000)
 
 	if got != want {
@@ -27,14 +27,14 @@ func TestPart2(t *testing.T) {
 }
 
 func BenchmarkPart1(b *testing.B) {
-	input := []int{1, 20, 8, 12, 0, 14}
+	input := []int32{1, 20, 8, 12, 0, 14}
 	for n := 0; n < b.N; n++ {
 		solve(input, 2020)
 	}
 }
 
 func BenchmarkPart2(b *testing.B) {
-	input := []int{1, 20, 8, 12, 0, 14}
+	input := []int32{1, 20, 8, 12, 0, 14}
 	for n := 0; n < b.N; n++ {
 		solve(input, 30000000)
 	}
